+++ src/yafowil/widget/dict/tests.py
@@ -5,20 +5,26 @@ from yafowil.base import factory
 from yafowil.compat import IS_PY2
 from yafowil.tests import fxml
 from yafowil.tests import YafowilTestCase
+import os
 import unittest
-import yafowil.loader  # noqa
 
 
 if not IS_PY2:
     from importlib import reload
 
 
+def np(path):
+    return path.replace('/', os.path.sep)
+
+
 class TestDictWidget(YafowilTestCase):
 
     def setUp(self):
         super(TestDictWidget, self).setUp()
+        from yafowil.widget import dict
         from yafowil.widget.dict import widget
         reload(widget)
+        dict.register()
 
     def test_empty_dict(self):
         # Create empty Dict widget
@@ -42,7 +48,7 @@ class TestDictWidget(YafowilTestCase):
             "    <class 'yafowil.base.Widget'>: body",
             ""
         ])
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <input class="hidden" id="input-mydict-exists"
                  name="mydict.exists" type="hidden" value="1"/>
@@ -148,7 +154,7 @@ class TestDictWidget(YafowilTestCase):
             "        <class 'yafowil.base.Widget'>: actions",
             ""
         ])
-        self.check_output("""
+        self.checkOutput("""
         <div>
           ...
           <tbody>
@@ -404,7 +410,7 @@ class TestDictWidget(YafowilTestCase):
             "        <class 'yafowil.base.Widget'>: value",
             ""
         ])
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <input class="hidden" id="input-mydict-exists"
                  name="mydict.exists" type="hidden" value="1"/>
@@ -520,6 +526,46 @@ class TestDictWidget(YafowilTestCase):
         expected = 'name="mydict.entry0.value" type="text" value="Value1"'
         self.assertTrue(rendered.find(expected) > -1)
 
+    def test_datatype_extraction(self):
+        widget = factory(
+            'error:dict',
+            name='mydict',
+            props={
+                'key_type': int,
+                'value_type': int
+            })
+        request = {
+            'mydict.exists': '1',
+            'mydict.entry0.key': '0',
+            'mydict.entry0.value': '1',
+        }
+        data = widget.extract(request=request)
+        self.assertFalse(data.has_errors)
+        self.assertEqual(
+            data.extracted,
+            odict([(0, 1)])
+        )
+        request = {
+            'mydict.exists': '1',
+            'mydict.entry0.key': 'a',
+            'mydict.entry0.value': '1',
+        }
+        data = widget.extract(request=request)
+        self.assertEqual(
+            data.errors,
+            [ExtractionError('Key a is not a valid integer.',)]
+        )
+        request = {
+            'mydict.exists': '1',
+            'mydict.entry0.key': '0',
+            'mydict.entry0.value': 'b',
+        }
+        data = widget.extract(request=request)
+        self.assertEqual(
+            data.errors,
+            [ExtractionError('Value b is not a valid integer.',)]
+        )
+
     def test_display_dict(self):
         widget = factory(
             'dict',
@@ -530,7 +576,7 @@ class TestDictWidget(YafowilTestCase):
                 'value_label': 'Value',
             },
             mode='display')
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <h5>Key: Value</h5>
           <dl>
@@ -551,7 +597,7 @@ class TestDictWidget(YafowilTestCase):
                 'value_label': 'Value'
             },
             mode='display')
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <h5>Key: Value</h5>
           <dl/>
@@ -567,7 +613,7 @@ class TestDictWidget(YafowilTestCase):
                 'value_label': lambda w, d: 'Computed Value'
             },
             mode='display')
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <h5>Computed Key: Computed Value</h5>
           <dl/>
@@ -585,7 +631,7 @@ class TestDictWidget(YafowilTestCase):
                 }
             },
             mode='display')
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <h5>B/C Key: B/C Value</h5>
           <dl/>
@@ -603,7 +649,7 @@ class TestDictWidget(YafowilTestCase):
                 }
             },
             mode='display')
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <h5>Computed B/C Key: Computed B/C Value</h5>
           <dl/>
@@ -616,11 +662,84 @@ class TestDictWidget(YafowilTestCase):
             name='display_dict',
             mode='display'
         )
-        self.check_output("""
+        self.checkOutput("""
         <div>
           <dl/>
         </div>
         """, fxml('<div>{}</div>'.format(widget())))
+
+    def test_resources(self):
+        factory.theme = 'default'
+        resources = factory.get_resources('yafowil.widget.dict')
+        self.assertTrue(resources.directory.endswith(np('/dict/resources')))
+        self.assertEqual(resources.name, 'yafowil.widget.dict')
+        self.assertEqual(resources.path, 'yafowil-dict')
+
+        scripts = resources.scripts
+        self.assertEqual(len(scripts), 1)
+
+        self.assertTrue(scripts[0].directory.endswith(np('/dict/resources')))
+        self.assertEqual(scripts[0].path, 'yafowil-dict')
+        self.assertEqual(scripts[0].file_name, 'widget.min.js')
+        self.assertTrue(os.path.exists(scripts[0].file_path))
+
+        styles = resources.styles
+        self.assertEqual(len(styles), 1)
+
+        self.assertTrue(
+            styles[0].directory.endswith(np('/dict/resources/default'))
+        )
+        self.assertEqual(styles[0].path, 'yafowil-dict/default')
+        self.assertEqual(styles[0].file_name, 'widget.css')
+        self.assertTrue(os.path.exists(styles[0].file_path))
+
+        factory.theme = 'bootstrap3'
+        resources = factory.get_resources('yafowil.widget.dict')
+        self.assertTrue(resources.directory.endswith(np('/dict/resources')))
+        self.assertEqual(resources.name, 'yafowil.widget.dict')
+        self.assertEqual(resources.path, 'yafowil-dict')
+
+        scripts = resources.scripts
+        self.assertEqual(len(scripts), 1)
+
+        self.assertTrue(scripts[0].directory.endswith(np('/dict/resources')))
+        self.assertEqual(scripts[0].path, 'yafowil-dict')
+        self.assertEqual(scripts[0].file_name, 'widget.min.js')
+        self.assertTrue(os.path.exists(scripts[0].file_path))
+
+        styles = resources.styles
+        self.assertEqual(len(styles), 1)
+
+        self.assertTrue(
+            styles[0].directory.endswith(np('/dict/resources/bootstrap'))
+        )
+        self.assertEqual(styles[0].path, 'yafowil-dict/bootstrap')
+        self.assertEqual(styles[0].file_name, 'widget.css')
+        self.assertTrue(os.path.exists(styles[0].file_path))
+
+        factory.theme = 'plone5'
+        resources = factory.get_resources('yafowil.widget.dict')
+        self.assertTrue(resources.directory.endswith(np('/dict/resources')))
+        self.assertEqual(resources.name, 'yafowil.widget.dict')
+        self.assertEqual(resources.path, 'yafowil-dict')
+
+        scripts = resources.scripts
+        self.assertEqual(len(scripts), 1)
+
+        self.assertTrue(scripts[0].directory.endswith(np('/dict/resources')))
+        self.assertEqual(scripts[0].path, 'yafowil-dict')
+        self.assertEqual(scripts[0].file_name, 'widget.min.js')
+        self.assertTrue(os.path.exists(scripts[0].file_path))
+
+        styles = resources.styles
+        self.assertEqual(len(styles), 1)
+
+        self.assertTrue(
+            styles[0].directory.endswith(np('/dict/resources/plone5'))
+        )
+        self.assertEqual(styles[0].path, 'yafowil-dict/plone5')
+        self.assertEqual(styles[0].file_name, 'widget.css')
+        self.assertTrue(os.path.exists(styles[0].file_path))
 
 
 if __name__ == '__main__':